fix: re-apply manual patches after regen
- Float→int type fixes for speaker/channel/num_words fields
- _sanitize_numeric_types helper in agent socket client
- construct_type keyword arg fix (generator uses positional, function requires keyword-only)
- Broad exception catch in socket clients (supports custom transports)
- Remove .bak files and update .fernignore

+++ src/deepgram/types/listen_v1response_results_channels_item_alternatives_item_paragraphs_paragraphs_item.py
@@ -14,8 +14,8 @@ class ListenV1ResponseResultsChannelsItemAlternativesItemParagraphsParagraphsIte
     sentences: typing.Optional[
         typing.List[ListenV1ResponseResultsChannelsItemAlternativesItemParagraphsParagraphsItemSentencesItem]
     ] = None
-    speaker: typing.Optional[float] = None
-    num_words: typing.Optional[float] = None
+    speaker: typing.Optional[int] = None
+    num_words: typing.Optional[int] = None
     start: typing.Optional[float] = None
     end: typing.Optional[float] = None
