+++ src/deepgram/types/listen_v1response_results_utterances_item_words_item.py
@@ -12,7 +12,7 @@ class ListenV1ResponseResultsUtterancesItemWordsItem(UncheckedBaseModel):
     start: typing.Optional[float] = None
     end: typing.Optional[float] = None
     confidence: typing.Optional[float] = None
-    speaker: typing.Optional[float] = None
+    speaker: typing.Optional[int] = None
     speaker_confidence: typing.Optional[float] = None
     punctuated_word: typing.Optional[str] = None
